src/main/python/scraper/requirements.txt
beautifulsoup4==4.13.4
requests==2.32.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0

